requirements.txt
streamlit
pandas
numpy
plotly
openai
python-dotenv
requests
pillow
folium
streamlit-folium
snowflake-connector-python[pandas]
snowflake-sqlalchemy
googlemaps
python-unsplash
scikit-learn
prophet
geopy
streamlit-option-menu
streamlit-aggrid
streamlit-plotly-events
